# openapi.yaml
openapi: 3.1.0
info:
  title: WebRetriever Local Control Plane
  version: 1.0.0
  description: Versioned contract for the loopback desktop sidecar. This document is authoritative before FastAPI implementation.
servers:
  - url: http://127.0.0.1:{port}
    variables:
      port:
        default: '43127'
security:
  - bearerAuth: []
paths:
...
  /api/v1/runtime/capabilities:
    get:
      operationId: getRuntimeCapabilities
      responses:
        '200':
          description: Supported local runtime limits and protocol version.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RuntimeCapabilities'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/account-profile:
    get:
      operationId: getAccountProfile
      responses:
        '200':
          description: The local desktop user's account profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
    put:
      operationId: updateAccountProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccountProfile'
      responses:
        '200':
          description: The saved local desktop user profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api/v1/model-services:
    get:
      operationId: listModelServices
      responses:
        '200':
          description: Configured model services without API keys.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ModelServiceSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: addModelService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelServiceInput'
      responses:
        '201':
          description: A model service was persisted; its API key is never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelServiceSummary'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api/v1/model-services/test:
    post:
      operationId: testUnsavedModelService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ModelServiceInput'
      responses:
        '200':
          description: A credential-safe result from one live provider request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelServiceTestResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api/v1/model-services/{service_name}/test:
    post:
      operationId: testSavedModelService
      parameters:
        - name: service_name
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 100
      responses:
        '200':
          description: A credential-safe result from testing a persisted service.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelServiceTestResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api/v1/run-preflights:
    post:
      operationId: preflightRun
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunSpec'
      responses:
        '200':
          description: Side-effect-free validation result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PreflightResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api/v1/task-submissions:
    post:
      operationId: submitTask
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskSubmissionRequest'
      responses:
        '202':
          description: A headed, single-task run was accepted asynchronously.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunAccepted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: A run is active or this idempotency key names a different request.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api/v1/runs:
    get:
      operationId: listRuns
      parameters:
        - $ref: '#/components/parameters/PageCursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Local run-history page.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      operationId: createRun
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RunSpec'
      responses:
        '202':
          description: Run was accepted and has begun asynchronously.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunAccepted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: A run is active or this idempotency key names a different request.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api/v1/runs/{run_id}:
    get:
      operationId: getRun
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Authoritative run snapshot.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RunSnapshot'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/RunNotFound'
  /api/v1/runs/{run_id}/cancel:
    post:
      operationId: cancelRun
      parameters:
        - $ref: '#/components/parameters/RunId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CancelRequest'
      responses:
        '200':
          description: The run was already terminal or cancellation was previously requested.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelResult'
        '202':
          description: Cooperative cancellation was accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CancelResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/RunNotFound'
        '422':
          $ref: '#/components/responses/ValidationFailed'
  /api/v1/runs/{run_id}/events:
    get:
      operationId: streamRunEvents
      parameters:
        - $ref: '#/components/parameters/RunId'
        - name: after
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
          description: Non-sensitive replay cursor; Last-Event-ID takes precedence.
        - name: Last-Event-ID
          in: header
          required: false
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Journal replay followed by live Server-Sent Events.
          content:
            text/event-stream:
              schema:
                type: string
                description: SSE id is RunEvent.event_id and data is a RunEvent JSON document.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/RunNotFound'
        '410':
          description: Requested cursor is older than retained journal events.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/EventsExpiredProblem'
  /api/v1/runs/{run_id}/artifacts:
    get:
      operationId: listArtifacts
      parameters:
        - $ref: '#/components/parameters/RunId'
        - $ref: '#/components/parameters/PageCursor'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: Opaque artifact identifiers under this run's canonical output root.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ArtifactPage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/RunNotFound'
  /api/v1/runs/{run_id}/artifacts/{artifact_id}:
    get:
      operationId: getArtifact
      parameters:
        - $ref: '#/components/parameters/RunId'
        - name: artifact_id
          in: path
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 128
      responses:
        '200':
          description: A sidecar-indexed artifact response. Range support is additive.
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/RunNotFound'
  /api/v1/control/shutdown:
    post:
      operationId: requestShutdown
      responses:
        '202':
          description: Sidecar entered draining shutdown coordination.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ShutdownAccepted'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    RunId:
      name: run_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    PageCursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
        minLength: 1
        maxLength: 256
    PageSize:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 50
  responses:
    Unauthorized:
      description: Missing, malformed, or incorrect launch bearer token.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    RunNotFound:
      description: No local run has this identifier.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    ValidationFailed:
      description: Request did not satisfy the stable v1 contract.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
  schemas:
    AccountProfile:
      type: object
      additionalProperties: false
      properties:
        name:
          type: string
          maxLength: 100
          default: 林晓宇
        email:
          type: string
          format: email
          maxLength: 254
          default: ''
        age:
          type: [integer, 'null']
          minimum: 0
          maximum: 150
          default: null
        work:
          type: string
          maxLength: 100
          default: ''
        organization:
          type: string
          maxLength: 100
          default: ''
    ModelServiceSummary:
      type: object
      additionalProperties: false
      required: [name, api_base, model, response_mode]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
          default: response
        api_base:
          type: string
          format: uri
          minLength: 1
          maxLength: 4096
        model:
          type: string
          minLength: 1
          maxLength: 256
          default: gpt-5.4
        response_mode:
          type: string
          enum: [responses, chat-completions]
          default: responses
    ModelServiceInput:
      allOf:
        - $ref: '#/components/schemas/ModelServiceSummary'
        - type: object
          required: [api_key]
          properties:
            api_key:
              type: string
              minLength: 1
              maxLength: 4096
              writeOnly: true
    ModelServiceTestResult:
      type: object
      additionalProperties: false
      required: [name, success, error_code]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        success:
          type: boolean
        error_code:
          type: [string, 'null']
...
    TaskSubmissionRequest:
      type: object
      additionalProperties: false
      required: [task, website_url]
      properties:
        task:
          type: string
          minLength: 1
          maxLength: 20000
        website_url:
          type: string
          minLength: 1
          maxLength: 4096
          format: uri
...
    RunSpec:
      type: object
      additionalProperties: false
      required: [schema_version, input_path, output_root, model, browser]
      properties:
        schema_version:
          const: 1
        input_path:
          type: string
          minLength: 1
          maxLength: 4096
        output_root:
          type: string
          minLength: 1
          maxLength: 4096
        project_url:
          type: [string, 'null']
          minLength: 1
          maxLength: 4096
          format: uri
        model:
          $ref: '#/components/schemas/CredentialProfileReference'
        browser:
          $ref: '#/components/schemas/LocalBrowserSpec'
        limits:
          $ref: '#/components/schemas/RunnerLimits'
        selection:
          $ref: '#/components/schemas/TaskSelection'
    RunAccepted:
      type: object
      additionalProperties: false
      required: [schema_version, run_id, status, created_at, snapshot_url, events_url]
      properties:
        schema_version:
          const: 1
        run_id:
          type: string
          format: uuid
        status:
          const: STARTING
        created_at:
          type: string
          format: date-time
        snapshot_url:
          type: string
        events_url:
          type: string
...
    RunSnapshot:
      type: object
      additionalProperties: false
      required: [schema_version, run_id, status, created_at, output_dir, last_event_id]
      properties:
        schema_version:
          const: 1
        run_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/RunStatus'
        created_at:
          type: string
          format: date-time
        started_at:
          type: [string, 'null']
          format: date-time
        finished_at:
          type: [string, 'null']
          format: date-time
        output_dir:
          type: string
        last_event_id:
          type: integer
          minimum: 0
        summary:
          type: [object, 'null']
        error:
          type: [object, 'null']
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, detail, instance, error_code]
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        instance:
          type: string
        error_code:
          type: string
          pattern: '^[a-z][a-z0-9_]*$'
        trace_id:
          type: [string, 'null']
        run_id:
          type: [string, 'null']
          format: uuid
        errors:
          type: [object, 'null']
          additionalProperties:
            type: array
            items:
              type: string
    EventsExpiredProblem:
      allOf:
        - $ref: '#/components/schemas/ProblemDetails'
        - type: object
          required: [minimum_event_id, snapshot_url]
          properties:
            minimum_event_id:
              type: integer
              minimum: 1
            snapshot_url:
              type: string
...
    RuntimeCapabilities:
      type: object
      additionalProperties: false
      required: [schema_version, browser_modes, limits]
      properties:
        schema_version:
          const: 1
        browser_modes:
          type: array
          const: [local]
        limits:
          $ref: '#/components/schemas/RunnerLimits'
    PreflightResult:
      type: object
      additionalProperties: false
      required: [schema_version, task_count, warnings]
      properties:
        schema_version:
          const: 1
        task_count:
          type: integer
          minimum: 1
          maximum: 100
        warnings:
          type: array
          items:
            type: string
    RunPage:
      type: object
      additionalProperties: false
      required: [schema_version, items, next_cursor]
      properties:
        schema_version:
          const: 1
        items:
          type: array
          items:
            $ref: '#/components/schemas/RunSnapshot'
        next_cursor:
          type: [string, 'null']
    CancelRequest:
      type: object
      additionalProperties: false
      properties:
        reason:
          type: string
          minLength: 1
          maxLength: 500
    CancelResult:
      type: object
      additionalProperties: false
      required: [schema_version, run_id, status, cancel_applied]
      properties:
        schema_version:
          const: 1
        run_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/RunStatus'
        cancel_applied:
          type: boolean
...
    ArtifactPage:
      type: object
      additionalProperties: false
      required: [schema_version, items, next_cursor]
      properties:
        schema_version:
          const: 1
        items:
          type: array
          items:
            $ref: '#/components/schemas/Artifact'
        next_cursor:
          type: [string, 'null']
...